// src/utils/enums/Tags.ts
export enum Tag {
  ACTION = 'Action',
  SPORTS = 'Sports',
  ADVENTURE = 'Adventure',
  SIMULATION = 'Simulation',
  RPG = 'RPG',
  FIGHT = 'Fight',
  FPS = 'FPS',
  PS5 = 'PS5',
  STEAM = 'Steam',
  SWITCH = 'SWITCH',
  XBOX = 'XBOX',
  WINDOWS = 'Windows',
  NUMBER = ''
}

// export enum Plataform {}